requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
msal==1.23.0
apscheduler==3.10.4
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
email-validator==2.1.0
pydantic-settings